More type conversion corrections with older gcc compilers.

+++ include/stxxl/bits/containers/priority_queue.h
@@ -814,7 +814,7 @@ void priority_queue<ConfigType>::dump_sizes() const
         STXXL_MSG("  grp " << i << " int" <<
                   " grpbuf=" << current_group_buffer_size(i) <<
                   " size=" << int_mergers[i].size() << "/" << capacity <<
-                  " (" << (int)(int_mergers[i].size() * 100.0 / capacity) << "%)" <<
+                  " (" << (int)((double)int_mergers[i].size() * 100.0 / (double)capacity) << "%)" <<
                   " space=" << int_mergers[i].is_space_available());
     }
     for (int i = 0; i < num_ext_groups; ++i) {
@@ -822,7 +822,7 @@ void priority_queue<ConfigType>::dump_sizes() const
         STXXL_MSG("  grp " << i + num_int_groups << " ext" <<
                   " grpbuf=" << current_group_buffer_size(i + num_int_groups) <<
                   " size=" << ext_mergers[i]->size() << "/" << capacity <<
-                  " (" << (int)(ext_mergers[i]->size() * 100.0 / capacity) << "%)" <<
+                  " (" << (int)((double)ext_mergers[i]->size() * 100.0 / (double)capacity) << "%)" <<
                   " space=" << ext_mergers[i]->is_space_available());
     }
     dump_params();
